Load neighbouring votes when show vote
To attempt to avoid having to load vote events when first show every
vote, load any unloaded neighbouring vote events when handle state
change for a vote. This works but does sometimes appear to slow the
chart transition for the current vote down, so may still be a better way
to do this.