Request neighbouring vote events when chart settles
Add an inbound port message sent when the bubble chart's alpha value
passes below a certain threshold, i.e. when the chart simulation is
mostly settled. This then triggers any needed requests for neighbouring
vote events.

This appears to improve how this works; if the requests for neighbouring
votes are made sooner it appears to noticeably interrupt and slow down
the simulation, while waiting for the simulation to fully complete
before doing this can take an unnecessarily long time.