Extract policies distinct from votes
Rather than loading votes multiple times for every policy they are
related to, now save policies in their own table and just add an
association to an also new join table to indicate every policy a vote is
related to.

This new information is not yet being used in the Elm app, but this has
been tweaked so it still correctly loads and uses the changed votes
data; it also now only shows every vote once due to these no longer
being duplicated.